Minor distrib tests refactor
