+++ tests/ignite/conftest.py
@@ -44,6 +44,10 @@ def distributed_context_single_node_nccl(local_rank):
 
     yield {'local_rank': local_rank}
 
+    # Perform some ops otherwise, next tests fail
+    device = "cuda"
+    _dummy_dist_op(device)
+
     dist.destroy_process_group()
 
 
@@ -64,6 +68,10 @@ def distributed_context_single_node_gloo(local_rank):
     dist.init_process_group(**dist_info)
 
     yield "distributed_context_single_node_gloo"
+
+    # Perform some ops otherwise, next tests fail
+    device = "cpu"
+    _dummy_dist_op(device)
 
     dist.destroy_process_group()
 
@@ -105,6 +113,10 @@ def distributed_context_multi_node_gloo(multi_node_conf):
 
     yield multi_node_conf
 
+    # Perform some ops otherwise, next tests fail
+    device = "cpu"
+    _dummy_dist_op(device)
+
     dist.destroy_process_group()
 
 
@@ -128,4 +140,21 @@ def distributed_context_multi_node_nccl(multi_node_conf):
 
     yield multi_node_conf
 
+    # Perform some ops otherwise, next tests fail
+    device = "cuda:{}".format(distributed_context_multi_node_nccl['local_rank'])
+    _dummy_dist_op(device)
+
     dist.destroy_process_group()
+
+
+def _dummy_dist_op(device):
+
+    def _gather(y):
+        output = [torch.zeros_like(y) for i in range(dist.get_world_size())]
+        dist.all_gather(output, y)
+        y = torch.cat(output, dim=0)
+        return y
+
+    y = torch.rand(10, 12, device=device)
+    y = _gather(y)
+    assert isinstance(y, torch.Tensor)
